+++ src/Enums/AlfColorNames.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Alf\Enums;
+
+enum AlfColorNames: string {
+
+    case BLACK = 'black';
+    case RED = 'red';
+    case GREEN = 'green';
+    case BLUE = 'blue';
+    case YELLOW = 'yellow';
+    case CYAN = 'cyan';
+    case MAGENTA = 'magenta';
+
+}